refactor(services): extract log_query SQL
Track D.9.  Lift query_logs (the dynamic SELECT plus its four
correlated cycle-aggregate subqueries plus the cycle-grouping
post-process) and summarize_rows out of routes/api/logs.py into
services/log_query.py.  The route keeps parser helpers + the HTMX-
shaped 422 partial.

The logs page handler (routes/pages.py) imported these helpers from
the sibling route module, which was a smell: a route reaching into
another route.  After this batch the page imports parser helpers
from routes/api/logs (still HTTP concerns) and SQL helpers from
services/log_query (data concerns) so the layering is honest.

Limit constants moved with the SQL: LIMIT_DEFAULT and LIMIT_MAX are
properties of the query (clamped at the boundary), the route imports
them back to keep the FastAPI Query default and upper bound aligned
with the service's clamp.  _LOAD_MORE_CHUNK_MAX stays
service-internal.

routes/api/logs.py also drops its own Jinja2Templates singleton in
favour of the centralised one from D.7; the local _get_templates
was missed in that batch.

Existing pinning tests in tests/test_routes/test_logs_pinning.py
keep their bodies; imports rebind the new public names through
``as`` aliases for compute_load_more_limit and summarize_rows so
the characterisation evidence carries over without touching the
asserts.  test_logs.py provides the route-level integration cover.

just quick green (1986 passed); mypy strict clean.
